refactor: move format function to helpers folder
