+++ src/components/DeckForm/TermInput.js
@@ -5,35 +5,15 @@ import { BiSolidImageAdd } from "react-icons/bi";
 import { TbWorldSearch } from "react-icons/tb";
 import { BsFillTrashFill } from "react-icons/bs";
 
-import "./TermInput.scss";
 import {
   WORDS_API_URL,
   WORDS_API_KEY,
   IMAGES_API_URL,
   IMAGES_API_KEY,
 } from "../../helpers/config";
+import { formatDescription } from "../../helpers/formatDescription";
 
-function formatDescription(data, term) {
-  if (!data) return "";
-
-  const description = data.shortdef.join(";\n");
-  const examples = data.def
-    .map((def) => def.sseq)
-    .flat(Infinity)
-    .filter((item) => typeof item !== "string" && item.dt)
-    .map((obj) => obj.dt)
-    .flat(Infinity)
-    .filter((item) => typeof item !== "string" && item.t)
-    .map((obj) =>
-      obj.t
-        .replaceAll(/{.*}(.*){.*}/g, "$1")
-        .replaceAll(term, "_".repeat(term.length))
-    )
-    .slice(0, 5)
-    .join("\n");
-
-  return `${description}${examples.length ? "\n".repeat(3) : ""}${examples}`;
-}
+import "./TermInput.scss";
 
 export default function TermInput({ field, fieldIndex, isOnlyItem, dispatch }) {
   const [loadedImages, setLoadedImages] = useState([]);
